Fix multiple customs lines in Stamp mapping.
Previously multiple CustomsLines tags would be generated if there was
more than one customs line, due to Savon's syntax. This changes the API
slightly, but fixes how multiple lines are passed since that would have
never worked. This is the new syntax to create a list of customs items

customs: {
  content_type: 'Merchandise',
  customs_lines: [
    {
      description: 'Some Product',
      quantity: 2,
      value: '10.0',
      country_of_origin: 'US',
      weight_lb: '13.3'
    },
    {
      description: 'Other Product',
      quantity: 2,
      value: '10.0',
      country_of_origin: 'US',
      weight_lb: '13.3'
    }
  ]
}